requirements.txt
asgiref==3.8.1
configparser==7.1.0
diff-match-patch==20241021
Django==5.1.6
django-cors-headers==4.6.0
django-crum==0.7.9
django-filter==24.3
django-import-export==4.3.3
djangorestframework==3.15.2
import-export==0.3.1
mysqlclient==2.2.6
python-dotenv==1.0.1
sqlparse==0.5.3
tablib==3.7.0
